Move can cast to python parser
